select length(name)
[git-p4: depot-paths = "//depot/qt/main/": change = 36830]

+++ tests/qdb/include/op.h
@@ -459,7 +459,7 @@ public:
     int exec( LocalSQLEnvironment* env )
     {
 	QString str = env->stack()->pop().toString();
-	env->stack()->push( str.length() );
+	env->stack()->push( (double) str.length() );
 	return 1;
     }
 };
